BSM-002 Remove port functionality
This commit removes the special handling of non-occupancy movements
involving ports, which was introduced as part of IBM-248.

As this is the only part of the system that needs to know what ports
are, the entire concept of ports is removed in this commit, with the
exception of database migrations.

It does *not* add a matching migration to drop any ports added to the
system, just in case we want to come back to them in future.