Optionally allow centre to be specified in URL
This commit allows clients to specify their centre in the URL as well
as the body, for example:

POST /irc_entry/ambridge/event

Clients may still specify their centre in the body only; this will be
accepted by the server.

Clients may in future specify their centre in the URL only; this will be
rejected by the server as it conflicts with the schema. If the schema
changes, the server should accept it without any changes beyond
informing the server of a new schema.

If the centre specified in the URL does not match the centre specified
in the body, the server will reject it with a HTTP status code of
422 Unprocessable Entity as defined in RFC 4918.